Remove druid implementation entirely
